p2p: Divide worker loop and peer struct methods
