+++ network/p2p/peer.go
@@ -1,15 +1,11 @@
 package p2p
 
 import (
-	"reflect"
 	"sync"
 	"time"
 
-	"github.com/airbloc/logger"
-
-	"github.com/airbloc/airbloc-go/network/p2p/message"
-
 	"github.com/airbloc/airbloc-go/network/p2p/handshake/identity"
+
 	"github.com/klaytn/klaytn/common"
 	"github.com/perlin-network/noise"
 	"github.com/pkg/errors"
@@ -75,79 +71,4 @@ func (peer Peer) messageWorkerWaitGroup() *sync.WaitGroup {
 
 func (peer Peer) messageAggregatorWaitGroup() *sync.WaitGroup {
 	return peer.Get(KeyPeerWaitGroupMessageAggregator).(*sync.WaitGroup)
-}
-
-func RunMessageAggregator(node Node, peer Peer) {
-	var (
-		aggregatorWaitGroup                            = peer.messageAggregatorWaitGroup()
-		aggregatedMessageChan chan<- aggregatedMessage = peer.aggregatedMessageChannnel()
-	)
-
-	aggregatorWaitGroup.Add(len(message.Opcodes))
-	for _, opcode := range message.Opcodes {
-		go func(opcode noise.Opcode) {
-			defer aggregatorWaitGroup.Done()
-			for {
-				select {
-				case <-peer.context().Done():
-					return
-				case receivedMessage := <-peer.Receive(opcode):
-					aggregatedMessageChan <- struct {
-						message noise.Message
-						opcode  noise.Opcode
-					}{
-						message: receivedMessage,
-						opcode:  opcode,
-					}
-				}
-			}
-		}(opcode)
-	}
-}
-
-func RunMessageWorker(node Node, peer Peer) {
-	var (
-		workerWaitgroup                                = peer.messageWorkerWaitGroup()
-		aggregatedMessageChan <-chan aggregatedMessage = peer.aggregatedMessageChannnel()
-	)
-
-	workerWaitgroup.Add(1)
-	go func() {
-		defer workerWaitgroup.Done()
-
-		for {
-			select {
-			case <-peer.context().Done():
-				return
-			case receivedMessage := <-aggregatedMessageChan:
-				node.logger().Debug("Message received", logger.Attrs{
-					"opcode": uint8(receivedMessage.opcode),
-					"type":   reflect.TypeOf(receivedMessage.message).String(),
-				})
-				handler, err := node.GetHandler(receivedMessage.opcode)
-				if err != nil {
-					node.logger().Error("Cannot get handler", logger.Attrs{
-						"error": logger.Attrs{
-							"type":    reflect.TypeOf(errors.Cause(err)).String(),
-							"message": err.Error(),
-						},
-						"message": reflect.TypeOf(receivedMessage.message).String(),
-					})
-					continue
-				}
-
-				err = handler(node.context(), receivedMessage.message, peer.Peer)
-				if err != nil {
-					node.logger().Error("Handler error occurred", logger.Attrs{
-						"error": logger.Attrs{
-							"type":    reflect.TypeOf(errors.Cause(err)).String(),
-							"message": err.Error(),
-						},
-						"message": reflect.TypeOf(receivedMessage.message).String(),
-					})
-					continue
-				}
-			}
-		}
-	}()
 }
